% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{estimate_elevation_dem}
\alias{estimate_elevation_dem}
\title{Estimate elevation data via Digital Elevation Model (DEM) from existing GPS data (.rds)}
\usage{
estimate_elevation_dem(rds_path, tif_dir)
}
\arguments{
\item{rds_path}{path of GPS data}

\item{tif_dir}{local directory path of DEM files (.tif)}
}
\value{
projection of elevation points on spatial points from GPS data
}
\description{
Estimate elevation data via Digital Elevation Model (DEM) from existing GPS data (.rds)
}
