% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{estimate_elevation_usgs}
\alias{estimate_elevation_usgs}
\title{Estimate elevation data with USGS API from existing GPS data (.rds)}
\usage{
estimate_elevation_usgs(rds_path)
}
\arguments{
\item{rds_path}{path of GPS data (.rds)}
}
\value{
elevation points based on GPS data
}
\description{
Estimate elevation data with USGS API from existing GPS data (.rds)
}
