% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Retrieve high resolution elevation data for the region of analysis from the internet}
\usage{
get_elevation(latmin, latmax, lonmin, lonmax)
}
\arguments{
\item{latmin}{minimum latitude for bounding box (degrees)}

\item{latmax}{maximum latitude for bounding box (degrees)}

\item{lonmin}{minimum longitude for bounding box (degrees)}

\item{lonmax}{maximum longitude for bounding box (degrees)}
}
\value{
elevation data as spatial points
}
\description{
Retrieve high resolution elevation data for the region of analysis from the internet
}
