% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny_cowtrackr}
\alias{run_shiny_cowtrackr}
\title{You can run the cowtrackr Shiny application by calling this function.}
\usage{
run_shiny_cowtrackr(rds_path)
}
\arguments{
\item{rds_path}{Path of cow data file to input}
}
\description{
You can run the cowtrackr Shiny application by calling this function.
}
